[#175] refactor : 폰트 추가로 공통 font.ts에 정리 및 해당 폰트 사용시 스타일 수정이 필요한 경우 구별하기 위해 사용할때 컴포넌트를 Text로 변경
